+++ hdate/__init__.py
@@ -7,7 +7,7 @@ of the Jewish calendrical date and times for a given location
 
 from hdate.date import HDate
 from hdate.hebrew_date import HebrewDate, Months
-from hdate.htables import HolidayTypes
+from hdate.holidays import HolidayTypes
 from hdate.location import Location
 from hdate.zmanim import Zmanim
